Make a prototype for indentation-sensitive method detection
This is necessary to correctly distinguish annotated class fields
preceded by a method declaration from type constructors inside that
method.

We have to duplicate all function-related contexts, but
the implementation is still incomplete: it does not understand braced
method definitions and suffers from abstract methods, which lack bodies.
Because it uses a dynamic regex (see `method-body` context), it slows
highlighting down by a factor of up to 2.8, depending on the source.

To sum up, I don't think it is worth it.